Move project / solution discovery to its own method
